feat: fixed proemtheus error not working on render
